Move Windows Service init out to seperate package
