+++ cmd/agent/initiate/entrypoint.go
@@ -1,4 +1,4 @@
-package main
+package initiate
 
 import (
 	"context"
